Improve backend code health
