+++ backend/internal/store/read_model_repository_ownership_test.go
@@ -1,0 +1,13 @@
+package store
+
+import (
+	"reflect"
+	"testing"
+)
+
+func TestReadModelRepositoryDoesNotKeepLegacyStoreBackReference(t *testing.T) {
+	repoType := reflect.TypeOf(pgReadModelRepository{})
+	if _, ok := repoType.FieldByName("legacy"); ok {
+		t.Fatal("pgReadModelRepository should receive explicit dependencies instead of a legacy *Store back-reference")
+	}
+}
